% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_group_plot.R
\name{box_group_plot}
\alias{box_group_plot}
\title{Plot a numerical (or categorical which can be converted to numerical) by a grouping variable.}
\usage{
box_group_plot(
  .df,
  xvar,
  xvar_name = xvar,
  yvar,
  yvar_name = yvar,
  dropna = TRUE,
  outliers = TRUE,
  reverse = FALSE,
  xmin = NA,
  xmax = NA,
  labels = NULL,
  sec.axis = TRUE,
  label_width = 20,
  coord_flip = FALSE,
  save = FALSE,
  bg = "transparent",
  width = 15.89,
  height = 10,
  units = "cm"
)
}
\arguments{
\item{.df}{A data.frame or tibble that contains the categorical variables in \code{var}.}

\item{xvar}{A numeric variable (or factor which can be converted to numeric) to plot on the x-axis.}

\item{xvar_name}{The x-axis name.}

\item{yvar}{A categorical variable to plot on the y-axis.}

\item{yvar_name}{The y-axis name.}

\item{dropna}{Plot missing grouping levels (i.e., NA groupings).}

\item{outliers}{Plot points representing outliers.}

\item{reverse}{Reverse the x-axis.}

\item{xmin}{Overwrite the minimum x-axis value if the value does not appear in the dataset.}

\item{xmax}{Overwrite the maximum x-axis value if the value does not appear in the dataset.}

\item{labels}{A named vector to provide the x-axis labels with human readable levels.}

\item{sec.axis}{A secondary axis to plot the numeric scores.}

\item{label_width}{The number of characters before words wrap. Used in the x- and y-axis labels.}

\item{coord_flip}{Have the boxplots vertical instead of horizontal.}

\item{save}{A logical to determine if the plot should be saved to your working directory "plots/**.png".}

\item{bg}{Background color.}

\item{width}{Width of plot.}

\item{height}{Height of plot.}

\item{units}{Unit of plot size.}
}
\value{
A ggplot object. This means you should be able to make further adjustments to it if required.
}
\description{
Plot a numerical (or categorical which can be converted to numerical) by a grouping variable.
}
\examples{
df <- data.frame(
  "record_id" = 1:50,
  "number" = sample(1:5, size = 50, replace = T),
  "group" = sample(as_factor(c("Group 1", "Group 2", "Group 3")), size = 50, replace = T)
)

box_group_plot(
  df,
  xvar = "number",
  xvar_name = "Numeric",
  yvar = "group",
  yvar_name = "Grouping",
  labels = c(
    "1" = "Strongly disagree",
    "2" = "Disagree",
    "3" = "Neutral",
    "4" = "Agree",
    "5" = "Strongly agree"
  )
)

}
