% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_box_plot.R
\name{consecutive_box_plot}
\alias{consecutive_box_plot}
\title{Plot a series of numerical variables in the form of a box plot.}
\usage{
consecutive_box_plot(
  .df,
  var,
  labels,
  label_width = 10,
  xlab = NULL,
  ylab = NULL,
  xmin = NA,
  xmax = NA,
  sec.axis = TRUE,
  reverse = FALSE,
  ordered_mean = FALSE,
  mean = TRUE,
  global_mean = TRUE,
  save = FALSE,
  bg = "transparent",
  width = 15.89,
  height = 10,
  units = "cm"
)
}
\arguments{
\item{.df}{A data.frame or tibble that contains the categorical variables in \code{var}.}

\item{var}{A vector of numerical variables to include in the plot. Can be a named list to automatically rename the variables in the plot.}

\item{labels}{A named vector to provide the x-axis labels with human readable levels.}

\item{label_width}{The number of characters before words wrap. Used in the x- and y-axis labels.}

\item{xlab}{The x-axis name.}

\item{ylab}{The y-axis name.}

\item{xmin}{Overwrite the minimum x-axis value if the value does not appear in the dataset.}

\item{xmax}{Overwrite the maximum x-axis value if the value does not appear in the dataset.}

\item{sec.axis}{A secondary axis to plot the numeric scores.}

\item{reverse}{Reverse the x-axis.}

\item{ordered_mean}{Order the \code{var} on the y-axis by mean}

\item{mean}{Display the mean of each variable}

\item{global_mean}{Display the global mean and median of all variables. Useful when displaying the results of all variables in a measure.}

\item{save}{A logical to determine if the plot should be saved to your working directory "plots/**.png".}

\item{bg}{Background color.}

\item{width}{Width of plot.}

\item{height}{Height of plot.}

\item{units}{Unit of plot size.}
}
\value{
A ggplot object. This means you should be able to make further adjustments to it if required.
}
\description{
Plot a series of numerical variables in the form of a box plot.
}
\examples{
set.seed(123)
df <- data.frame(
  "number1" = sample(1:5, size = 50, replace = TRUE),
  "number2" = sample(1:5, size = 50, replace = TRUE),
  "number3" = sample(1:5, size = 50, replace = TRUE)
)

consecutive_box_plot(
  df,
  var = c(
    "Number 1" = "number1",
    "Number 2" = "number2",
    "Number 3" = "number3"
  ),
  labels = c(
    "1" = "Strongly diagree",
    "2" = "Disagree",
    "3" = "Neutral",
    "4" = "Agree",
    "5" = "Strongly agree"
  ),
  ordered_mean = TRUE
)


}
