% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_cat_plot.R
\name{consecutive_cat_plot}
\alias{consecutive_cat_plot}
\title{Plot a series of categorical variables.}
\usage{
consecutive_cat_plot(
  .df,
  var,
  colors,
  y_lab = TRUE,
  print_data = FALSE,
  rev_fill = FALSE,
  text_size = NULL,
  pct_label = T,
  xaxis_size = 9,
  yaxis_size = 9,
  round_digits = 1,
  legend_size = 12,
  label_width = 14,
  pct_cut = 0.15,
  save = FALSE,
  bg = "transparent",
  width = 15.89,
  height = 16,
  units = "cm"
)
}
\arguments{
\item{.df}{A data.frame or tibble that contains the categorical variables in \code{var}.}

\item{var}{A vector of categorical variables to include in the plot. Can be a named list to automatically rename the variables in the plot.}

\item{colors}{A named vector of colours to assign to each level of the \code{var} variable in format of \verb{description == color code}. The order of the vector will be used in the legend.}

\item{y_lab}{A logical indicating if the y-axis labels should be displayed.}

\item{print_data}{A helper option to print some raw data to the console. Can be helpful when debugging, or adding data to reports.}

\item{rev_fill}{Used to reverse the color of the bars in the plot.}

\item{text_size}{The size of the text in the labels.}

\item{pct_label}{To show the \% sign or not.}

\item{xaxis_size}{The size of the text on the x-axis.}

\item{yaxis_size}{The size of the text on the y-axis.}

\item{round_digits}{Number of digits to round to.}

\item{legend_size}{The size of the text on the legend.}

\item{label_width}{The number of characters before words wrap. Used in the y-axis and legend.}

\item{pct_cut}{Categories with a pct > than this value will not print the \% on the plot.}

\item{save}{A logical to determine if the plot should be saved to your working directory "plots/**.png".}

\item{bg}{Background color.}

\item{width}{Width of plot.}

\item{height}{Height of plot.}

\item{units}{Unit of plot size.}
}
\value{
A ggplot object. This means you should be able to make further adjustments to it if required.
}
\description{
Plot a series of categorical variables.
}
\examples{
set.seed(123)
df <- data.frame(
  "apples" = sample(forcats::as_factor(c("Good", "Neutral", "Bad")), size = 50, replace = TRUE),
  "bananas" = sample(forcats::as_factor(c("Good", "Neutral", "Bad")), size = 50, replace = TRUE),
  "pears" = sample(forcats::as_factor(c("Good", "Neutral", "Bad")), size = 50, replace = TRUE)
)

consecutive_cat_plot(
  df,
  c("Apples" = "apples", "Bananas" = "bananas", "Pears" = "pears"),
  colors = c(
    "Bad" = "#b44218",
    "Neutral" = "#b4b4b4",
    "Good" = "#179dab"
  )
)
}
