% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factored_case_match.R
\name{factored_case_match}
\alias{factored_case_match}
\title{Automatic factoring of dplyr::case_match}
\usage{
factored_case_match(.x, ..., .default = NULL)
}
\arguments{
\item{.x}{A vector to match against}

\item{...}{A sequence of two sided formulas: \code{old_values ~ new_value}. Can either be predefined and reused by use of \code{!!!lvl} called each time.}

\item{.default}{The value used when values in .x aren't matched by any of the LHS inputs. If NULL, the default, a missing value will be used.}
}
\value{
A vector of type factor with the levels in the factor in the order they were put in the list of formulas.
}
\description{
Automatic factoring of dplyr::case_match
}
\examples{
df <- data.frame(
  a = c(1, 3, NA, NA, 5, 2),
  b = c(2, 3, NA, 3, 1, 4)
)

lvl <- rlang::exprs(
  4 ~ "Large",
  3 ~ "Medium",
  2 ~ "Small",
  1 ~ "Very small"
)

dplyr::mutate(
df,
    a = factored_case_match(
      a,
      !!!lvl
    ),
    b = factored_case_match(
      b,
1 ~ "Very small",
2 ~ "Small",
3 ~ "Medium",
      4 ~ "Large"
    )
  )
}
