% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factored_case_when.R
\name{factored_case_when}
\alias{factored_case_when}
\title{Automatic factoring of dplyr::case_when}
\usage{
factored_case_when(..., .default = NULL)
}
\arguments{
\item{...}{A sequence of two sided formulas: \code{logical_test ~ new_value}.}

\item{.default}{The value used when values in .x aren't matched by any of the LHS inputs. If NULL, the default, a missing value will be used.}
}
\value{
A vector of type factor with the levels in the factor in the order they were put in the list of formulas.
}
\description{
Automatic factoring of dplyr::case_when
}
\examples{
df <- data.frame(
  x = c(1, 2, 3, 4, NA),
  y = c(NA, 5, 4, 3, 2)
)

dplyr::mutate(
df,
    x_label = factored_case_when(
      x == 4 ~ "A pig",
      x == 3 ~ "Medium",
      x == 2 ~ "Three",
      x == 1 ~ "Very small",
      is.na(x) ~ "Missing"
    ),
   y_label = factored_case_when(
y >= 4 ~ "High",
.default = "Low"
)
  )
}
