% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_group_plot.R
\name{cat_group_plot}
\alias{cat_group_plot}
\title{Plot a categorical variable over a grouping variable.}
\usage{
cat_group_plot(
  .df,
  xvar,
  xlab = NULL,
  yvar,
  ylab = NULL,
  colors,
  label_width = 20,
  pct_cut = 0.06,
  legend_size = 12,
  save = FALSE,
  bg = "transparent",
  width = 15.89,
  height = 10,
  units = "cm"
)
}
\arguments{
\item{.df}{A data.frame or tibble that contains a categorical and grouping variable.}

\item{xvar}{The categorical variable to include in the plot.}

\item{xlab}{The x-axis label.}

\item{yvar}{The grouping variable to include in the plot.}

\item{ylab}{The y-axis label.}

\item{colors}{A named vector of colours to assign to each level of the \code{xvar} variable in format of \verb{description == color code}. The order of the vector will be used in the legend.}

\item{label_width}{The number of characters before words wrap. Used in the y-axis and legend.}

\item{pct_cut}{Categories with a pct > than this value will not print the \% on the plot.}

\item{legend_size}{The size of the text on the legend.}

\item{save}{A logical to determine if the plot should be saved to your working directory "plots/**.png".}

\item{bg}{Background color.}

\item{width}{Width of plot.}

\item{height}{Height of plot.}

\item{units}{Unit of plot size.}
}
\value{
A ggplot object. This means you should be able to make further adjustments to it if required.
}
\description{
Plot a categorical variable over a grouping variable.
}
\examples{
df <- data.frame(
  "apples" = sample(
    forcats::as_factor(c("Good", "Neutral", "Bad")),
    size = 50,
    replace = TRUE
  ),
  "group" = sample(
    forcats::as_factor(c("Group 1", "Group 2", "Group 3")),
    size = 50,
    replace = TRUE
  )
)
cat_group_plot(
  df,
  xvar = "apples",
  yvar = "group",
  colors = c(
    "Bad" = "#b44218",
    "Neutral" = "#b4b4b4",
    "Good" = "#179dab"
  )
)
}
